Server: keep a history list
This is accessible via the "history" command, just as the
queue is. You can use this to check what's been playing
lately.

The history grows without bound. In practice, this isn't a
big deal. The average track is a few minutes long and has an
absolute path of about 50 characters or so. At that rate, we
waste a whopping 23K per day of constant playing. In only
2-3 years of constant playing, we'll match the 17 megabyte
bloat we get just from loading perl and gstreamer into
memory.